# spring-swagger-codegen-employer/src/main/java/com/example/employer/api/FindemployerDetailsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.8).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.example.employer.api;

import com.example.employer.model.Employer;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2019-06-18T13:26:19.982+03:00[Europe/Helsinki]")
@Api(value = "findemployerDetails", description = "the findemployerDetails API")
public interface FindemployerDetailsApi {

    @ApiOperation(value = "Find employer by ID", nickname = "getemployerDetails", notes = "Returns a single employer", response = Employer.class, tags={ "employer", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Employer.class),
        @ApiResponse(code = 400, message = "Invalid employer ID supplied"),
        @ApiResponse(code = 404, message = "employer not found") })
    @RequestMapping(value = "/findemployerDetails/{employerId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Employer> getemployerDetails(@ApiParam(value = "ID of employer to return",required=true) @PathVariable("employerId") Long employerId);

}